Fix issue #144 - show filename for ReST errors
Issue #144: reStructuredText rendering errors don't show filename,
just "string".  The actual ReST renderer,
docutils.core.publish_parts(), supports a "source_path" parameter,
which can be used to set the filename of the rendered file.  With
this paramter set, in case of error, the filename will be shown
instead of just "<string>".

This fix fundamentally modifies the structure of the
wok/renderers.py module.  Previously, all render() functions were
defined as classmethods.  This change makes them instance methods.
Each page object gets assigned its own Renderer instance.  For the
ReStructuredText renderer, an optional source_path parameter is
allowed by the constructor.  If set, this will be stored with the
instance, and passed on to the docutils function.

A new function is introduced in wok/renderers.py: factory().  This
is called in wok/engine.py when the renderer is assigned to the
page.  This does the work of looking at all available renderers, and
matching the filename extention to the appropriate renderer, and
returning a valid object instance.